move time formatter to framework
