+++ src/no/runsafe/UserControl/command/Played.java
@@ -1,12 +1,12 @@
 package no.runsafe.UserControl.command;
 
-import no.runsafe.UserControl.TimeFormatter;
 import no.runsafe.UserControl.database.PlayerSessionLog;
 import no.runsafe.framework.api.IScheduler;
 import no.runsafe.framework.api.command.IBranchingExecution;
 import no.runsafe.framework.api.command.argument.IArgumentList;
 import no.runsafe.framework.api.command.player.PlayerAsyncCommand;
 import no.runsafe.framework.api.player.IPlayer;
+import no.runsafe.framework.tools.TimeFormatter;
 
 import java.time.Duration;
